Update README to include goals and release index
Include goals and intentions in README to allow for evaluation at the
end of the project.

Release html boilerplate index as landing page for the site. Already
completed the very basic structure, with a view to next adding three
individual recipe pages and further formatting landing page so it isn't
so dire and bleak.