Add first recipe page and test folder push
The barely-formatted first recipe page will be added to the repo,
but mainly this commit is to see how the folder push displays on GitHub
so I can better my understanding of the git push command.

Next plan is to complete formatting of the page including image, lists,
and link back home.